# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Mavi API
  version: 0.1.0
  description: |
    JSON CRUD for candidates, their profiles, roles and matches. Every request
    outside `/health` names a persona with the `X-Role` header (`talent`,
    `employer` or `ops`) and may identify itself with `X-Actor` (the candidate
    id for talent, an email for ops). A missing or unknown role is a 401; a
    known role calling an operation it may not use is a 403.

    `PUT` on candidates and roles replaces the fields you send and keeps the
    rest; an explicit `null` clears an optional field and is a 422 on a
    required one. `PUT …/profile` replaces the whole profile. Taxonomy fields
    (certifications, software) accept free text and are stored as canonical
    ids from infra/taxonomy.json; an unknown value is a 422.
servers:
  - url: http://localhost:8080
security:
  - role: []
...
paths:
...
  /roles:
    post:
      tags: [roles]
      operationId: createRole
      summary: Create a role
      x-roles: [employer, ops]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Role" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    get:
      tags: [roles]
      operationId: listRoles
      summary: List roles
      description: Talent only ever receives open roles; a `status` filter other than `open` returns an empty list for talent.
      x-roles: [talent, employer, ops]
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/RoleStatus" }
        - name: company
          in: query
          description: Exact match.
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Role" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /roles/intake:
    post:
      tags: [roles]
      operationId: intakeRole
      summary: Create a role from a pasted job description and queue its matching run
      description: |
        Employer intake. The AI service's `/parse-jd` reads the job description
        in the request and the role is stored with the raw text as
        `description`, the whole extraction as `requirements`, and the
        must-haves, nice-to-haves and hard-filter requirements in their own
        fields, so the response shows the employer what was understood. The
        role is embedded before the answer (from the structured requirements,
        through `/embed-batch`, as an `embed_role` job would), and a
        `match_role` job is queued for it: the matching run, which reranks the
        candidates who pass the role's filters and writes them to the role's
        matches, the best of them as `pending_review` for ops. None of them is
        visible to the employer until ops releases it.

        The parse is a chat-model call, so the request can take a while. If
        only the embedding fails the role is still created: `embedded_at` is
        `null` in the response and an `embed_role` job fills it in later.
        Correct anything the parser got wrong with `PUT /roles/{id}`.
      x-roles: [employer, ops]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleIntakeInput" }
      responses:
        "201":
          description: The role as stored, and its matching job.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleIntake" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422":
          description: |
            `description` is missing, blank or longer than the parser accepts; or the job
            description names no title and none was sent (`fields.title`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: The AI service, which parses the job description, is unavailable or its model's output did not validate. Nothing was stored; try again.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /roles/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [roles]
      operationId: getRole
      summary: Fetch a role
      description: A role that is not open is a 404 for talent.
      x-roles: [talent, employer, ops]
      responses:
        "200":
          description: The role.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Role" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [roles]
      operationId: updateRole
      summary: Update a role
      description: Replaces the fields sent and keeps the rest. Changing the description clears the stored embedding.
      x-roles: [employer, ops]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Role" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [roles]
      operationId: deleteRole
      summary: Delete a role
      x-roles: [employer, ops]
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /jobs:
    post:
      tags: [jobs]
      operationId: createJob
      summary: Enqueue a background job
      description: |
        Adds a job to the Postgres-backed queue. `kind` must be one the worker
        has a handler for (`embed_role`, `embed_profile`, `parse_resume`); anything else is a
        422 naming the known kinds. If an identical job (same `kind` and
        `payload`) is already queued, that job is returned with a 200 instead
        of adding a duplicate. The worker picks the job up within its poll
        interval; poll `GET /jobs/{id}` until `status` is `succeeded` or `failed`.
      x-roles: [ops]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobCreate" }
      responses:
        "201":
          description: Queued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "200":
          description: An identical job was already queued; it is returned instead of a duplicate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    get:
      tags: [jobs]
      operationId: listJobs
      summary: List background jobs
      x-roles: [ops]
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/JobStatus" }
        - name: kind
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    role:
      type: apiKey
      in: header
      name: X-Role
      description: A `Persona` (`talent`, `employer` or `ops`). The `mavi_role` cookie is a session fallback.
    actor:
      type: apiKey
      in: header
      name: X-Actor
      description: >-
        The candidate id for talent (required for talent to read anything), an email for ops (required for
        the review operations, whose audit trail records it).
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
    Limit:
      name: limit
      in: query
      description: Page size; default 50, max 200.
      schema: { type: integer, minimum: 1, maximum: 200 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0 }
  responses:
    BadRequest:
      description: Malformed JSON or an unknown field.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or unknown `X-Role`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: >-
        The role may not call this operation, a talent request is missing its candidate id, or a review
        operation is missing its reviewer (`X-Actor`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such row, or one the caller may not see.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: >-
        A unique constraint (duplicate email, duplicate role/candidate pair), a delete blocked by history, or a
        review operation the role's matches do not allow as they stand (`error` says why; nothing was changed).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: One or more fields are invalid; `fields` maps each to a message.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Id:
      type: string
      description: A UUID, as text.
      example: 11111111-0000-0000-0000-000000000001
    JobId:
      type: integer
      format: int64
      description: A job's id. Jobs are numbered from a sequence, not UUIDs.
      example: 42
    CalendarDate:
      type: string
      format: date
      description: A calendar day, `YYYY-MM-DD`.
      example: "2026-10-14"
      x-go-type: Date
    JSONObject:
      type: object
      description: A free-form JSON object stored as JSONB.
      additionalProperties: true
      x-go-type: RawJSON
...
    RoleStatus:
      type: string
      enum: [open, filled, closed]
...
    JobStatus:
      type: string
      description: "`queued` → `running` → `succeeded`, or `running` → `queued` (retry) → … → `failed`."
      enum: [queued, running, succeeded, failed]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: validation failed
        fields:
          type: object
          description: Per-field messages on a 422.
          additionalProperties: { type: string }
...
    Role:
      type: object
      required:
        [id, title, company, description, requirements, must_haves, nice_to_haves, required_certifications,
         required_software, min_years_experience, timezone, min_overlap_hours, hours_per_week, starts_on, status,
         embedding_model, embedded_at, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        title: { type: string }
        company: { type: string, nullable: true }
        description:
          type: string
          description: The raw job description.
        requirements:
          allOf: [{ $ref: "#/components/schemas/JSONObject" }]
          description: The parser's structured extraction (the AI service's `RoleRequirements`).
        must_haves:
          type: array
          items: { type: string }
        nice_to_haves:
          type: array
          items: { type: string }
        required_certifications:
          type: array
          description: Canonical taxonomy ids.
          items: { type: string }
        required_software:
          type: array
          description: Canonical taxonomy ids.
          items: { type: string }
        min_years_experience:
          type: integer
          nullable: true
          minimum: 0
          maximum: 70
          description: Fewest years of experience the role accepts; `null` when the JD gives no number (a `0` sent in is stored as `null`).
        timezone:
          type: string
          nullable: true
          description: IANA zone the role operates in. Its working day is 09:00 to 17:00 there.
        min_overlap_hours:
          type: integer
          nullable: true
          minimum: 1
          maximum: 8
          description: Hours of the role's working day a candidate's working hours must cover; `null` when the role does not ask.
        hours_per_week:
          type: integer
          nullable: true
          minimum: 1
          maximum: 80
          description: Hours a week the role needs; a candidate must offer at least as many. `null` when the role does not ask.
        starts_on:
          allOf: [{ $ref: "#/components/schemas/CalendarDate" }]
          nullable: true
        status: { $ref: "#/components/schemas/RoleStatus" }
        embedding_model: { type: string, nullable: true }
        embedded_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    RoleInput:
      type: object
      description: |
        Body for creating or updating a role. Every field is optional on the
        wire so a `PUT` can send only what changes; `title` is required when
        creating and may not be nulled.
      properties:
        title: { type: string }
        company: { type: string, nullable: true }
        description:
          type: string
          nullable: true
          description: "`null` clears it to an empty string."
        requirements:
          allOf: [{ $ref: "#/components/schemas/JSONObject" }]
          x-go-type-skip-optional-pointer: true
        must_haves:
          type: array
          items: { type: string }
          x-go-type-skip-optional-pointer: true
        nice_to_haves:
          type: array
          items: { type: string }
          x-go-type-skip-optional-pointer: true
        required_certifications:
          type: array
          items: { type: string }
          x-go-type-skip-optional-pointer: true
        required_software:
          type: array
          items: { type: string }
          x-go-type-skip-optional-pointer: true
        min_years_experience: { type: integer, nullable: true, minimum: 0, maximum: 70 }
        timezone: { type: string, nullable: true }
        min_overlap_hours:
          type: integer
          nullable: true
          minimum: 0
          maximum: 8
          description: Needs `timezone`. Supply it at intake when the JD does not state it; a `0` is stored as `null`.
        hours_per_week:
          type: integer
          nullable: true
          minimum: 0
          maximum: 80
          description: Supply it at intake when the JD does not state it; a `0` is stored as `null`.
        starts_on:
          allOf: [{ $ref: "#/components/schemas/CalendarDate" }]
          nullable: true
        status:
          allOf: [{ $ref: "#/components/schemas/RoleStatus" }]
          description: Defaults to `open` on create.
    RoleIntakeInput:
      type: object
      required: [description]
      properties:
        description:
          type: string
          description: The job description as pasted, up to 60,000 characters.
        title:
          type: string
          nullable: true
          description: Overrides the title the parser finds; required when the job description names none.
        company:
          type: string
          nullable: true
          description: Overrides the company the parser finds.
    RoleIntake:
      type: object
      required: [role, matching_job]
      properties:
        role: { $ref: "#/components/schemas/Role" }
        matching_job:
          allOf: [{ $ref: "#/components/schemas/Job" }]
          description: The `match_role` job queued for the role.

    # -- matches --------------------------------------------------------------
...
    Job:
      type: object
      description: |
        One background job in the Postgres-backed queue (the `jobs` table). A
        worker claims a `queued` job whose `run_at` has passed, runs the handler
        for its `kind`, and marks it `succeeded`. A failing attempt puts it back
        to `queued` with `run_at` pushed out by a backoff until `attempts`
        reaches `max_attempts`, when it lands in `failed` with `last_error` set.
      required:
        [id, kind, payload, status, priority, run_at, attempts, max_attempts, last_error, worker, started_at,
         finished_at, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/JobId" }
        kind:
          type: string
          description: Names the handler that runs it, e.g. `embed_role`, `parse_resume` or `match_role`.
        payload:
          allOf: [{ $ref: "#/components/schemas/JSONObject" }]
          description: 'Handler-specific input, e.g. `{"role_id": "…"}`.'
        status: { $ref: "#/components/schemas/JobStatus" }
        priority:
          type: integer
          description: Higher runs first.
        run_at:
          type: string
          format: date-time
          description: Not before this time. Retry backoff pushes it into the future.
        attempts:
          type: integer
          description: Claims so far, including one in progress.
        max_attempts:
          type: integer
          minimum: 1
        last_error:
          type: string
          nullable: true
          description: The most recent failed attempt's error; kept on a `failed` job, cleared on success.
        worker:
          type: string
          nullable: true
          description: Id of the worker that last claimed it.
        started_at:
          type: string
          format: date-time
          nullable: true
          description: When the current or most recent attempt was claimed.
        finished_at:
          type: string
          format: date-time
          nullable: true
          description: Set when the job reached `succeeded` or `failed`.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    JobCreate:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          description: Must be a kind the worker has a handler for; otherwise a 422 naming the known kinds.
          example: embed_role
        payload:
          allOf: [{ $ref: "#/components/schemas/JSONObject" }]
          x-go-type-skip-optional-pointer: true
        priority:
          type: integer
          minimum: -32768
          maximum: 32767
          description: Defaults to 0. Higher runs first.
        run_at:
          type: string
          format: date-time
          description: Delay the job until this time. Defaults to now.
        max_attempts:
          type: integer
          minimum: 1
          maximum: 20
          description: Defaults to 3.